requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
tifffile
Pillow
tqdm
